вынес логики спавна предмета в отдельный класс, вынес функции gathervalidateteams в класс настройки команд, выделение языковых функций в utility
