+++ game/scripts/vscripts/GameMode.lua
@@ -3,7 +3,7 @@ local __TS__Class = ____lualib.__TS__Class
 local __TS__New = ____lualib.__TS__New
 local __TS__Decorate = ____lualib.__TS__Decorate
 local __TS__SourceMapTraceBack = ____lualib.__TS__SourceMapTraceBack
-__TS__SourceMapTraceBack(debug.getinfo(1).short_src, {["8"] = 1,["9"] = 1,["10"] = 2,["11"] = 2,["12"] = 3,["13"] = 3,["14"] = 4,["15"] = 4,["16"] = 6,["17"] = 15,["18"] = 16,["19"] = 15,["21"] = 18,["22"] = 19,["23"] = 82,["24"] = 87,["25"] = 87,["26"] = 87,["27"] = 87,["28"] = 87,["29"] = 88,["30"] = 88,["31"] = 88,["32"] = 88,["33"] = 88,["34"] = 89,["35"] = 89,["36"] = 89,["37"] = 89,["38"] = 89,["39"] = 92,["40"] = 92,["41"] = 92,["42"] = 93,["43"] = 96,["44"] = 97,["45"] = 105,["46"] = 106,["47"] = 107,["49"] = 92,["50"] = 92,["51"] = 81,["52"] = 21,["53"] = 23,["54"] = 24,["55"] = 26,["56"] = 29,["57"] = 30,["58"] = 31,["59"] = 34,["60"] = 35,["61"] = 36,["62"] = 37,["63"] = 38,["64"] = 39,["65"] = 40,["66"] = 41,["67"] = 42,["68"] = 43,["69"] = 46,["70"] = 47,["71"] = 48,["72"] = 49,["73"] = 52,["74"] = 53,["75"] = 56,["76"] = 59,["77"] = 60,["78"] = 21,["79"] = 63,["80"] = 65,["81"] = 63,["82"] = 71,["84"] = 73,["85"] = 73,["86"] = 74,["87"] = 73,["90"] = 77,["91"] = 79,["92"] = 71,["93"] = 115,["94"] = 116,["95"] = 118,["96"] = 121,["97"] = 122,["98"] = 123,["99"] = 124,["100"] = 127,["101"] = 128,["102"] = 129,["103"] = 129,["104"] = 129,["105"] = 129,["106"] = 129,["107"] = 129,["108"] = 130,["109"] = 132,["110"] = 135,["111"] = 136,["112"] = 138,["113"] = 115,["114"] = 142,["115"] = 143,["116"] = 142,["117"] = 146,["118"] = 147,["119"] = 149,["120"] = 161,["122"] = 173,["123"] = 175,["124"] = 175,["125"] = 175,["126"] = 175,["128"] = 146,["129"] = 179,["130"] = 180,["131"] = 179,["132"] = 186,["133"] = 187,["134"] = 186,["135"] = 192,["136"] = 194,["137"] = 196,["138"] = 197,["139"] = 199,["142"] = 192,["143"] = 205,["144"] = 206,["145"] = 207,["146"] = 208,["147"] = 209,["148"] = 210,["149"] = 211,["150"] = 212,["151"] = 213,["152"] = 214,["153"] = 216,["154"] = 217,["155"] = 218,["156"] = 219,["157"] = 221,["158"] = 222,["159"] = 223,["160"] = 225,["161"] = 226,["162"] = 229,["163"] = 231,["164"] = 231,["165"] = 231,["166"] = 231,["167"] = 231,["168"] = 231,["169"] = 231,["170"] = 231,["171"] = 231,["172"] = 231,["173"] = 205,["174"] = 15,["175"] = 16});
+__TS__SourceMapTraceBack(debug.getinfo(1).short_src, {["8"] = 1,["9"] = 1,["10"] = 2,["11"] = 2,["12"] = 3,["13"] = 3,["14"] = 4,["15"] = 4,["16"] = 6,["17"] = 15,["18"] = 16,["19"] = 15,["21"] = 18,["22"] = 19,["23"] = 19,["24"] = 19,["25"] = 19,["26"] = 82,["27"] = 87,["28"] = 87,["29"] = 87,["30"] = 87,["31"] = 87,["32"] = 88,["33"] = 88,["34"] = 88,["35"] = 88,["36"] = 88,["37"] = 89,["38"] = 89,["39"] = 89,["40"] = 89,["41"] = 89,["42"] = 92,["43"] = 92,["44"] = 92,["45"] = 93,["46"] = 96,["47"] = 97,["48"] = 105,["49"] = 106,["50"] = 107,["52"] = 92,["53"] = 92,["54"] = 81,["55"] = 21,["56"] = 23,["57"] = 24,["58"] = 26,["59"] = 29,["60"] = 30,["61"] = 31,["62"] = 34,["63"] = 35,["64"] = 36,["65"] = 37,["66"] = 38,["67"] = 39,["68"] = 40,["69"] = 41,["70"] = 42,["71"] = 43,["72"] = 46,["73"] = 47,["74"] = 48,["75"] = 49,["76"] = 52,["77"] = 53,["78"] = 56,["79"] = 59,["80"] = 60,["81"] = 21,["82"] = 63,["83"] = 65,["84"] = 63,["85"] = 71,["87"] = 73,["88"] = 73,["89"] = 74,["90"] = 73,["93"] = 77,["94"] = 79,["95"] = 71,["96"] = 115,["97"] = 116,["98"] = 118,["99"] = 121,["100"] = 122,["101"] = 123,["102"] = 124,["103"] = 127,["104"] = 128,["105"] = 129,["106"] = 129,["107"] = 129,["108"] = 129,["109"] = 129,["110"] = 129,["111"] = 130,["112"] = 132,["113"] = 135,["114"] = 136,["115"] = 115,["116"] = 142,["117"] = 143,["118"] = 142,["119"] = 146,["120"] = 147,["121"] = 149,["122"] = 161,["124"] = 173,["125"] = 175,["126"] = 175,["127"] = 175,["128"] = 175,["130"] = 146,["131"] = 179,["132"] = 180,["133"] = 179,["134"] = 186,["135"] = 187,["136"] = 186,["137"] = 192,["138"] = 194,["139"] = 196,["140"] = 197,["141"] = 199,["144"] = 192,["145"] = 205,["146"] = 206,["147"] = 207,["148"] = 208,["149"] = 209,["150"] = 210,["151"] = 211,["152"] = 212,["153"] = 213,["154"] = 214,["155"] = 216,["156"] = 217,["157"] = 218,["158"] = 219,["159"] = 221,["160"] = 222,["161"] = 223,["162"] = 225,["163"] = 226,["164"] = 229,["165"] = 231,["166"] = 231,["167"] = 231,["168"] = 231,["169"] = 231,["170"] = 231,["171"] = 231,["172"] = 231,["173"] = 231,["174"] = 231,["175"] = 205,["176"] = 15,["177"] = 16});
 local ____exports = {}
 local ____tstl_2Dutils = require("lib.tstl-utils")
 local reloadable = ____tstl_2Dutils.reloadable
@@ -19,7 +19,10 @@ local GameMode = ____exports.GameMode
 GameMode.name = "GameMode"
 function GameMode.prototype.____constructor(self)
     self.TeamsSettings = __TS__New(TeamsSettings)
-    self.Overthrow = __TS__New(Overthrow)
+    self.Overthrow = __TS__New(
+        Overthrow,
+        self.TeamsSettings:GatherAndRegisterValidTeams()
+    )
     self:configure()
     ListenToGameEvent(
         "game_rules_state_change",
@@ -109,7 +112,6 @@ function GameMode.prototype.configure(self)
     self:settingGameModeEntity()
     GameRules:SetShowcaseTime(2)
     GameRules:SetHeroSelectionTime(heroSelectionTime)
-    self.Overthrow:GatherAndRegisterValidTeams()
 end
 function GameMode.prototype.SpawnGoldEntity(self, spawnPoint)
     self.Overthrow:SpawnGoldEntity(spawnPoint)
