+++ src/vscripts/OverthrowEvents.ts
@@ -1,0 +1,77 @@
+export function ExecuteOrderFilter( filterTable : ExecuteOrderFilterEvent ) : boolean{
+	/*
+	for k, v in pairs( filterTable ) {
+		print("EO: " .. k .. " " .. tostring(v) )
+	}
+	*/
+
+    const orderType = filterTable.order_type 
+    if ( orderType != UnitOrder.PICKUP_ITEM || filterTable.issuer_player_id_const == -1 ){
+        return true
+    }
+    else {
+        const item = EntIndexToHScript(filterTable.entindex_target) as CDOTA_Item_Physical
+        if (item == undefined) {
+            return true
+        }
+        const pickedItem = item.GetContainedItem()
+
+        //print(pickedItem.GetAbilityName())
+        if (pickedItem == undefined) {
+            return true
+        }
+        switch(pickedItem.GetAbilityName()){
+            case "item_bag_of_gold":{
+                const hero = EntIndexToHScript(filterTable.units["0"]) as CDOTA_BaseNPC_Hero
+                if(hero == undefined || !hero.IsRealHero())
+                    return false;
+                break;
+            }
+            case "item_treasure_chest":{
+                const player = PlayerResource.GetPlayer(filterTable.issuer_player_id_const)
+                if(player==undefined){
+                    return true
+                }
+                const hero = player.GetAssignedHero()
+                
+                // determine if we can scoop the neutral or not
+                // we need either a free backpack slot or a free neutral item slot
+                let bAllowPickup = false
+                const hNeutralItem = hero.GetItemInSlot(16) // InventorySlot.NEUTRAL_SLOT -- не работает почему-то, на практике ровно nil а не 16
+                if (hNeutralItem == undefined) {
+                    bAllowPickup = true
+                    //print( '^^^Empty neutral slot!' )
+                }
+                else {
+                    let numBackpackItems = 0
+                    for (let nItemSlot = 0; nItemSlot < DOTA_ITEM_INVENTORY_SIZE; nItemSlot++) {
+                        const hItem = hero.GetItemInSlot( nItemSlot ) // index zero based
+                        if (hItem && hItem.IsInBackpack()) {
+                            numBackpackItems = numBackpackItems + 1
+                        }
+                    }
+                    //print( '^^^Backpack slots = ' .. numBackpackItems )
+                    if (numBackpackItems < 3) {
+                        bAllowPickup = true
+                    }
+                }		
+
+                if (bAllowPickup) {
+                    //print("inventory has space")
+                    return true
+                }
+                else {
+                    //print("Moving to target instead")
+                    const position = item.GetAbsOrigin()
+                    filterTable.position_x = position.x
+                    filterTable.position_y = position.y
+                    filterTable.position_z = position.z
+                    filterTable.order_type = UnitOrder.MOVE_TO_POSITION 
+                    return true
+                }
+                break;
+            }
+        }
+    }
+    return true
+}
